requirements.txt
openai
requests
python-dotenv
tencentcloud-sdk-python
zhipuai
spark_ai_python
websocket-client
